+++ src/routes/index.ts
@@ -1,11 +1,9 @@
 import express from 'express';
-import helloRouter from './routeHello';
-import postRouter from './routePost';
+import postRouter from './postRouter';
 import userRouter from './userRouter';
 
 const router = express.Router();
 
-router.use('/hello', helloRouter);
 router.use('/posts', postRouter);
 router.use('/auth', userRouter);
